+++ R/f_12_hour.R
@@ -1,0 +1,38 @@
+#' Format 12 Hour Times
+#'
+#' Format times to the typical 12 hour '%I:%M %p' in-text format.
+#'
+#' @param x A vector of coercable times.
+#' @param format A character string specifying the time output format.
+#' @param pad.char A character to use for leading padding if lengths of output
+#' are unequal.
+#' @param \ldots Other arguments passed to \code{\link[base]{as.POSIXct}}.
+#' @return Returns a string of publication ready 12 hour time stamps.
+#' @export
+#' @rdname f_12_hour
+#' @examples
+#' f_12_hour(Sys.time())
+#' f_12_hour(Sys.time(), pad.char ='0')
+#' f_12_hour(Sys.time(), pad.char =' ')
+#' f_12_hour(Sys.time(), '%I:%M:%S %p')
+#' set.seed(10)
+#' times <- as.POSIXct(sample(seq_len(1e4), 12), origin = '1970-01-01')
+#' paste(f_12_hour(range(times)), collapse = ' to ')
+f_12_hour <- function(x, format = '%I:%M %p', pad.char = '', ...){
+
+    out <- format(as.POSIXct(x, ...), format = format)
+    gsub('^0', pad.char, out)
+
+}
+
+
+#' @export
+#' @include utils.R
+#' @rdname f_12_hour
+ff_12_hour <- functionize(f_12_hour)
+
+
+
+
+
+