+++ R/f_12_hour.R
@@ -15,15 +15,38 @@
 #' f_12_hour(Sys.time(), pad.char ='0')
 #' f_12_hour(Sys.time(), pad.char =' ')
 #' f_12_hour(Sys.time(), '%I:%M:%S %p')
+#' f_12_hour(0:24, '%I %p')
 #' set.seed(10)
 #' times <- as.POSIXct(sample(seq_len(1e4), 12), origin = '1970-01-01')
 #' paste(f_12_hour(range(times)), collapse = ' to ')
-f_12_hour <- function(x, format = '%I:%M %p', pad.char = '', ...){
+f_12_hour <- function(x = Sys.time(), format = '%I:%M %p', pad.char = '', ...){
+    UseMethod('f_12_hour')
+}
+
+
+#' @export
+#' @rdname f_12_hour
+#' @method f_12_hour default
+f_12_hour.default <- function(x, format = '%I:%M %p', pad.char = '', ...){
 
     out <- format(as.POSIXct(x, ...), format = format)
     gsub('^0', pad.char, out)
 
 }
+
+
+#' @export
+#' @rdname f_12_hour
+#' @method f_12_hour integer
+f_12_hour.integer <- function(x, format = '%I:%M %p', pad.char = '', ...){
+
+
+    out <- format(as.POSIXct(paste0("2017-08-18 ", ifelse(nchar(x) == 1, '0', ''), x, ":00:00")), format=format)
+
+    gsub('^0', pad.char, out)
+
+}
+
 
 
 #' @export