+++ R/constants.R
@@ -1,0 +1,32 @@
+#' Constants
+#'
+#' \code{constant_months}A constant for ordered month names.
+#'
+#' @keywords datasets
+#' @name constant_months
+#' @rdname constants
+#' @export
+constant_months <- month.name
+
+#' Constants
+#'
+#' \code{constant_months_abbreviation} - A constant for ordered month abbreviations.
+#'
+#' @keywords datasets
+#' @name constant_months_abbreviation
+#' @rdname constants
+#' @export
+constant_months_abbreviation <- month.abb
+
+#' Constants
+#'
+#' \code{constant_weekdays} - A constant for ordered weekdays.
+#'
+#' @keywords datasets
+#' @name constant_weekdays
+#' @rdname constants
+#' @export
+constant_weekdays <- c("Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday")
+
+
+