+++ R/f_12_hour.R
@@ -15,7 +15,7 @@
 #' f_12_hour(Sys.time(), pad.char ='0')
 #' f_12_hour(Sys.time(), pad.char =' ')
 #' f_12_hour(Sys.time(), '%I:%M:%S %p')
-#' f_12_hour(0:24, '%I %p')
+#' f_12_hour(c(NA, 0:24), '%I %p')
 #' set.seed(10)
 #' times <- as.POSIXct(sample(seq_len(1e4), 12), origin = '1970-01-01')
 #' paste(f_12_hour(range(times)), collapse = ' to ')
@@ -29,8 +29,14 @@ f_12_hour <- function(x = Sys.time(), format = '%I:%M %p', pad.char = '', ...){
 #' @method f_12_hour default
 f_12_hour.default <- function(x, format = '%I:%M %p', pad.char = '', ...){
 
-    out <- format(as.POSIXct(x, ..., origin = "1960-01-01"), format = format)
-    gsub('^0', pad.char, out)
+    x <- as.POSIXct(x, ..., origin = "1960-01-01")
+    nalocs <- is.na(x)
+    x[nalocs] <- as.POSIXct(Sys.Date())
+
+    out <- format(x, format = format)
+    out <- gsub('^0', pad.char, out)
+    out[nalocs] <- NA
+    out
 
 }
 
@@ -40,10 +46,14 @@ f_12_hour.default <- function(x, format = '%I:%M %p', pad.char = '', ...){
 #' @method f_12_hour integer
 f_12_hour.integer <- function(x, format = '%I:%M %p', pad.char = '', ...){
 
+    nalocs <- is.na(x)
+    x[nalocs] <- 1
 
     out <- format(as.POSIXct(paste0("2017-08-18 ", ifelse(nchar(x) == 1, '0', ''), x, ":00:00"), origin = "1960-01-01"), format=format)
 
-    gsub('^0', pad.char, out)
+    out <- gsub('^0', pad.char, out)
+    out[nalocs] <- NA
+    out
 
 }
 
@@ -54,9 +64,14 @@ f_12_hour.integer <- function(x, format = '%I:%M %p', pad.char = '', ...){
 f_12_hour.numeric <- function(x, format = '%I:%M %p', pad.char = '', ...){
 
     x <- as.integer(x)
+    nalocs <- is.na(x)
+    x[nalocs] <- 1
+
     out <- format(as.POSIXct(paste0("2017-08-18 ", ifelse(nchar(x) == 1, '0', ''), x, ":00:00"), origin = "1960-01-01"), format=format)
 
-    gsub('^0', pad.char, out)
+    out <- gsub('^0', pad.char, out)
+    out[nalocs] <- NA
+    out
 
 }